refactor: merge gateway and auth verifier
