+++ Cargo.toml
@@ -14,7 +14,6 @@ rust-version = "1.76.0"
 soroban-sdk = { version = "21.0.1-preview.3" }
 axelar-soroban-interfaces = { version = "^0.1.0", path = "contracts/axelar-soroban-interfaces" }
 axelar-soroban-std = { version = "^0.1.0", path = "packages/axelar-soroban-std" }
-axelar-auth-verifier = { version = "^0.1.0", path = "contracts/axelar-auth-verifier" }
 axelar-gas-service = { version = "^0.1.0", path = "contracts/axelar-gas-service" }
 axelar-gateway = { version = "^0.1.0", path = "contracts/axelar-gateway" }
 axelar-operators = { version = "^0.1.0", path = "contracts/axelar-operators" }
