--- contracts/axelar-auth-verifier/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "axelar-auth-verifier"
-version = "0.1.0"
-edition = { workspace = true }
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-axelar-soroban-interfaces = { workspace = true }
-axelar-soroban-std = { workspace = true }
-ed25519-dalek = { version = "^2.1", default-features = false, optional = true }
-rand = { version = "^0.8.5", optional = true }
-sha3 = { version = "*", optional = true }
-soroban-sdk = { workspace = true }
-
-[dev_dependencies]
-axelar-soroban-std = { workspace = true, features = ["testutils"] }
-ed25519-dalek = { version = "^2.1" }
-sha3 = { version = "*" }
-soroban-sdk = { workspace = true, features = ["testutils"] }
-rand = { version = "^0.8.5" }
-
-[features]
-testutils = ["soroban-sdk/testutils", "axelar-soroban-std/testutils", "ed25519-dalek", "sha3", "rand"]
