+++ contracts/axelar-gateway/src/error.rs
@@ -8,4 +8,14 @@ pub enum Error {
     EmptyMessages = 2,
     RotationAlreadyExecuted = 3,
     NotLatestSigners = 4,
+    InvalidOperators = 5,
+    InvalidThreshold = 6,
+    DuplicateOperators = 7,
+    MalformedSigners = 8,
+    LowSignaturesWeight = 9,
+    InvalidProof = 10,
+    InvalidSigners = 11,
+    InsufficientRotationDelay = 12,
+    InvalidSignatures = 13,
+    InvalidWeights = 14,
 }
