+++ contracts/axelar-soroban-interfaces/src/axelar_gateway.rs
@@ -5,8 +5,15 @@ use crate::types::{Message, Proof, WeightedSigners};
 /// Interface for the Axelar Gateway.
 #[contractclient(name = "AxelarGatewayClient")]
 pub trait AxelarGatewayInterface {
-    /// Initialize the gateway with the given auth module address.
-    fn initialize(env: Env, auth_module: Address, operator: Address);
+    /// Initialize the gateway
+    fn initialize(
+        env: Env,
+        operator: Address,
+        domain_separator: BytesN<32>,
+        previous_signer_retention: u64,
+        minimum_rotation_delay: u64,
+        initial_signers: Vec<WeightedSigners>
+    );
 
     /// Call a contract on another chain with the given payload. The destination address can validate the contract call on the destination gateway.
     fn call_contract(
@@ -16,6 +23,8 @@ pub trait AxelarGatewayInterface {
         destination_address: String,
         payload: Bytes,
     );
+
+    fn approve_messages(env: Env, messages: Vec<Message>, proof: Proof);
 
     /// Validate if a contract call with the given payload BytesN<32> and source caller info is approved,
     /// preventing re-validation (i.e distinct contract calls can be validated at most once).
@@ -42,7 +51,10 @@ pub trait AxelarGatewayInterface {
     /// Return true if a contract call with the given payload BytesN<32> and source caller info has been executed.
     fn is_message_executed(env: Env, message_id: String, source_chain: String) -> bool;
 
-    fn approve_messages(env: Env, messages: Vec<Message>, proof: Proof);
-
     fn rotate_signers(env: Env, signers: WeightedSigners, proof: Proof);
+
+    fn transfer_operatorship(env: Env, new_operator: Address);
+
+    fn operator(env: &Env) -> Address;
+
 }
