+++ contracts/axelar-soroban-interfaces/src/lib.rs
@@ -2,7 +2,6 @@
 
 pub mod types;
 
-pub mod axelar_auth_verifier;
 pub mod axelar_executable;
 pub mod axelar_gas_service;
 pub mod axelar_gateway;
