+++ integration-tests/Cargo.toml
@@ -13,5 +13,4 @@ crate-type = ["rlib"]
 soroban-sdk = { workspace = true, features = ["testutils"] }
 axelar-soroban-std = { workspace = true, features = ["testutils"] }
 axelar-soroban-interfaces = { workspace = true }
-axelar-auth-verifier = { workspace = true, features = [ "testutils" ] }
 axelar-gateway = { workspace = true, features = [ "testutils" ] }
